BUG : data access at the same time
to generate sampleExtraction vector, we have to merge some points vectors.
However, this merge results is the same in case of SAR and optical sensor use separatly.
A check 'if file exists' is not enough to fix the issue. A simple fix is done by adding
a random sleep before checking if the file exists (a few time is needed to write the merged file).
Also create a entire step can be done, if this solution is not the good one.